Extract interface for notification repository
